# docs/api.yaml
openapi: 3.0.3
info:
  title: Book API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server

paths:
  /books:
    post:
      summary: Create a new book
      tags:
        - books
      description: Create a new book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostBookRequest'
      responses:
        '201':
          description: Book created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List books
      tags:
        - books
      description: List books
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  books:
                    type: array
                    items:
                      $ref: '#/components/schemas/Book'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


components:
  schemas:
    PostBookRequest:
      type: object
      required:
        - title
        - author
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 256
          x-oapi-codegen-extra-tags:
            validate: "required,min=1,max=256"
        author:
          type: string
          minLength: 1
          maxLength: 256
          x-oapi-codegen-extra-tags:
            validate: "required,min=1,max=256"
        description:
          type: string
          maxLength: 1000
          x-oapi-codegen-extra-tags:
            validate: "max=1000"
          
    Book:
      type: object
      required:
        - title
        - author
        - description
      properties:
        title:
          type: string
        author:
          type: string
        description:
          type: string

    ErrorResponse:
      type: object
      required:
        - message
        - status
      properties:
        status:
          type: integer
        message:
          type: string