# internal/gen/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
)

// Book defines model for Book.
type Book struct {
	Author      string `json:"author"`
	Description string `json:"description"`
	Title       string `json:"title"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Message string `json:"message"`
	Status  int    `json:"status"`
}

// PostBookRequest defines model for PostBookRequest.
type PostBookRequest struct {
	Author      string  `json:"author" validate:"required,min=1,max=256"`
	Description *string `json:"description,omitempty" validate:"max=1000"`
	Title       string  `json:"title" validate:"required,min=1,max=256"`
}

// PostBooksJSONRequestBody defines body for PostBooks for application/json ContentType.
type PostBooksJSONRequestBody = PostBookRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List books
	// (GET /books)
	GetBooks(ctx echo.Context) error
	// Create a new book
	// (POST /books)
	PostBooks(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetBooks converts echo context to params.
func (w *ServerInterfaceWrapper) GetBooks(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetBooks(ctx)
	return err
}

// PostBooks converts echo context to params.
func (w *ServerInterfaceWrapper) PostBooks(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostBooks(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(baseURL+"/books", wrapper.GetBooks)
	router.POST(baseURL+"/books", wrapper.PostBooks)

}

type GetBooksRequestObject struct {
}

type GetBooksResponseObject interface {
	VisitGetBooksResponse(w http.ResponseWriter) error
}

type GetBooks200JSONResponse struct {
	Books *[]Book `json:"books,omitempty"`
}

func (response GetBooks200JSONResponse) VisitGetBooksResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type GetBooks400JSONResponse ErrorResponse

func (response GetBooks400JSONResponse) VisitGetBooksResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type GetBooks500JSONResponse ErrorResponse

func (response GetBooks500JSONResponse) VisitGetBooksResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type PostBooksRequestObject struct {
	Body *PostBooksJSONRequestBody
}

type PostBooksResponseObject interface {
	VisitPostBooksResponse(w http.ResponseWriter) error
}

type PostBooks201JSONResponse Book

func (response PostBooks201JSONResponse) VisitPostBooksResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type PostBooks400JSONResponse ErrorResponse

func (response PostBooks400JSONResponse) VisitPostBooksResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type PostBooks500JSONResponse ErrorResponse

func (response PostBooks500JSONResponse) VisitPostBooksResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// List books
	// (GET /books)
	GetBooks(ctx context.Context, request GetBooksRequestObject) (GetBooksResponseObject, error)
	// Create a new book
	// (POST /books)
	PostBooks(ctx context.Context, request PostBooksRequestObject) (PostBooksResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// GetBooks operation middleware
func (sh *strictHandler) GetBooks(ctx echo.Context) error {
	var request GetBooksRequestObject

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetBooks(ctx.Request().Context(), request.(GetBooksRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetBooks")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetBooksResponseObject); ok {
		return validResponse.VisitGetBooksResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// PostBooks operation middleware
func (sh *strictHandler) PostBooks(ctx echo.Context) error {
	var request PostBooksRequestObject

	var body PostBooksJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostBooks(ctx.Request().Context(), request.(PostBooksRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostBooks")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PostBooksResponseObject); ok {
		return validResponse.VisitPostBooksResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}